Added amends to connection handler to open and close the db gracefully after every request

+++ src/main/java/com/test/learningdi/db/DbApi.java
@@ -14,29 +14,36 @@ public class DbApi implements DbCallbacks {
     @NotNull
     private Map<Integer, Employee> employeeDatabase;
     private Integer autoIncrement = -1;
+    private boolean isOpen = false;
 
     public DbApi(){
         employeeDatabase = new HashMap<>();
         autoIncrement++;
     }
 
+
     @Override
     public int add(Employee employeeObject) {
         synchronized (this) {
-            if (employeeExists(employeeObject)) {
-                return DbResponseConstants.ALREADY_EXISTS;
+            if(isOpen) {
+                if (employeeExists(employeeObject)) {
+                    return DbResponseConstants.ALREADY_EXISTS;
+                }
+                employeeDatabase.put(autoIncrement++, employeeObject);
+                return DbResponseConstants.ADDED_SUCCESS;
             }
-            employeeDatabase.put(autoIncrement++, employeeObject);
-            return DbResponseConstants.ADDED_SUCCESS;
+            return DbResponseConstants.DB_CLOSED;
         }
     }
 
     private boolean employeeExists(Employee obj) {
-        if(employeeDatabase != null){
-            for(Map.Entry employee : employeeDatabase.entrySet()){
-                Employee employee1 = (Employee)employee.getValue();
-                if(employee1.getName().contains(obj.getName())){
-                    return true;
+        if(isOpen) {
+            if (employeeDatabase != null) {
+                for (Map.Entry employee : employeeDatabase.entrySet()) {
+                    Employee employee1 = (Employee) employee.getValue();
+                    if (employee1.getName().contains(obj.getName())) {
+                        return true;
+                    }
                 }
             }
         }
@@ -51,44 +58,53 @@ public class DbApi implements DbCallbacks {
     @Override
     public int remove(int id) {
         synchronized (this) {
-            if (employeeDatabase != null) {
-                if (employeeDatabase.get(id) != null) {
-                    if (employeeDatabase.remove(id) != null) {
-                        return DbResponseConstants.DELETED_SUCCESS;
+            if(isOpen) {
+                if (employeeDatabase != null) {
+                    if (employeeDatabase.get(id) != null) {
+                        if (employeeDatabase.remove(id) != null) {
+                            return DbResponseConstants.DELETED_SUCCESS;
+                        }
                     }
+                    return DbResponseConstants.FAILED;
                 }
             }
+            return DbResponseConstants.DB_CLOSED;
         }
-        return DbResponseConstants.FAILED;
     }
 
     @Override
     public int remove(Employee object) {
         synchronized (this) {
-            if (employeeDatabase != null) {
-                for (Map.Entry employee : employeeDatabase.entrySet()) {
-                    Employee employee1 = (Employee) employee.getValue();
-                    if (employee1.getName().contains(object.getName())) {
-                        employeeDatabase.remove(employee1);
-                        return DbResponseConstants.DELETED_SUCCESS;
+            if (isOpen) {
+                if (employeeDatabase != null) {
+                    for (Map.Entry employee : employeeDatabase.entrySet()) {
+                        Employee employee1 = (Employee) employee.getValue();
+                        if (employee1.getName().contains(object.getName())) {
+                            employeeDatabase.remove(employee1);
+                            return DbResponseConstants.DELETED_SUCCESS;
+                        }
                     }
                 }
+                return DbResponseConstants.FAILED;
             }
         }
-        return DbResponseConstants.FAILED;
+        return DbResponseConstants.DB_CLOSED;
     }
 
     @Override
     public int update(int id, Employee object) {
         synchronized (this) {
-            if (employeeDatabase != null) {
-               if(employeeDatabase.get(id) != null){
-                   employeeDatabase.put(id, object);
-                   return DbResponseConstants.UPDATED_SUCCESS;
-               }
+            if (isOpen) {
+                if (employeeDatabase != null) {
+                    if (employeeDatabase.get(id) != null) {
+                        employeeDatabase.put(id, object);
+                        return DbResponseConstants.UPDATED_SUCCESS;
+                    }
+                }
+                return DbResponseConstants.FAILED;
             }
-            return DbResponseConstants.FAILED;
         }
+        return DbResponseConstants.DB_CLOSED;
     }
 
     @Override
@@ -105,10 +121,33 @@ public class DbApi implements DbCallbacks {
 
     @Override
     public Employee get(int id){
-        if(employeeDatabase != null){
-            employeeDatabase.get(id);
+        synchronized (this) {
+            if (isOpen) {
+                if (employeeDatabase != null) {
+                    employeeDatabase.get(id);
+                }
+                return null;
+            }
         }
         return null;
+    }
+
+    @Override
+    public int clear() {
+        this.employeeDatabase.clear();
+        return -1;
+    }
+
+    @Override
+    public int close() {
+        isOpen = false;
+        return -1;
+    }
+
+    @Override
+    public int open() {
+        isOpen = true;
+        return 0;
     }
 
 }
